backend/requirement.txt
absl-py==2.3.1
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
astunparse==1.6.3
attrs==25.3.0
bcrypt==3.2.0
beautifulsoup4==4.13.5
ccxt==4.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.6
curl_cffi==0.13.0
cycler==0.12.1
ecdsa==0.19.1
fastapi==0.116.1
flatbuffers==25.2.10
fonttools==4.59.2
frozendict==2.4.6
frozenlist==1.7.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
idna==3.10
joblib==1.5.2
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
multidict==6.6.4
multitasking==0.0.12
namex==0.1.0
numpy==2.3.2
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
passlib==1.7.4
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.32.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycares==4.10.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2025.2
requests==2.32.5
rich==14.1.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==1.4.23
starlette==0.47.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.15.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
xgboost==3.0.4
yarl==1.20.1
yfinance==0.2.65
